starting to streamline socket io
