+++ config/sequelize.js
