+++ controllers/index.js
@@ -3,8 +3,6 @@ const router = require("express").Router();
 const api = require("./api");
 router.use("/api", api);
 
-const chat = require('./socket/chat')
-router.use('/chat', chat)
 
 // router.get("/", (req, res) => res.json("home"));
