--- controllers/socket/chat.js
@@ -1,5 +1,0 @@
-const router = require('express').Router()
-
-router.get('/', (req, res)=>res.json('chat endpoint'))
-
-module.exports = router
