+++ models/Users.js
@@ -1,5 +1,5 @@
 const { Model, DataTypes } = require('sequelize');
-const sequelize = require('../config');
+const sequelize = require('../config/sequelize');
 const bcrypt = require("bcrypt")
 
 class User extends Model {}
